requirements.txt
aiohttp==3.6.2
Pillow==7.1.1
pytest==5.4.1
pytest-asyncio==0.10.0
redis==3.4.1